+++ CONTRIBUTING.md
@@ -1,6 +1,6 @@
 ## Contributing
 
-The files to modify are under the `src` folder. `src/assets` are JavaScript files written in es6 that get compiled
+The files to modify are under the `src` folder. `src/assets` are JavaScript files written in JavaScript that get compiled
 through babel into `lib/sitemapper.js`.
 
 ### Build
@@ -11,7 +11,7 @@ To build the `lib` directory with the compiled assets use this command
 npm run build
 ```
 
-This uses [Babel](http://babeljs.io/) to compile the files. Make sure to run `npm run build` before submitting a pull request.
+This uses [Babel](http://babeljs.io/) to compile the files. The prepack step will run `npm run build` when submitting a pull request.
 
 ```bash
 # Run examples/index.js
@@ -58,7 +58,10 @@ src/
   assets/
     sitemapper.js
   examples/
+    google.js
     index.js
   tests/
     test.js
+    test.ts.ts
+    tsconfig.json
 ```
